Avoid an int cast of double in vbr_rate_correction
In the calculation of frame_window, avoid the int cast of
cpi->ppi->twopass.stats_buf_ctx->total_stats->count (a double) because
in theory it could be greater than INT_MAX. Use the int cast on the
result of AOMMIN(16, ...), which is <= 16

This is equivalent to the first change in the libvpx CL
https://chromium-review.googlesource.com/c/webm/libvpx/+/5670633.

+++ av1/encoder/ratectrl.c
@@ -2598,12 +2598,12 @@ static void vbr_rate_correction(AV1_COMP *cpi, int *this_frame_target) {
 #endif
   int64_t frame_target = *this_frame_target;
 
-  const int stats_count =
+  const double stats_count =
       cpi->ppi->twopass.stats_buf_ctx->total_stats != NULL
-          ? (int)cpi->ppi->twopass.stats_buf_ctx->total_stats->count
-          : 0;
-  const int frame_window = AOMMIN(
-      16, (int)(stats_count - (int)cpi->common.current_frame.frame_number));
+          ? cpi->ppi->twopass.stats_buf_ctx->total_stats->count
+          : 0.0;
+  const int frame_window =
+      (int)AOMMIN(16, stats_count - cpi->common.current_frame.frame_number);
   assert(VBR_PCT_ADJUSTMENT_LIMIT <= 100);
   if (frame_window > 0) {
     const int64_t max_delta =
